+++ server/index.ts
@@ -1,6 +1,7 @@
 import register from "./register";
 import bootstrap from "./bootstrap";
 import destroy from "./destroy";
+import config from "./config";
 import routes from "./routes";
 import controllers from "./controllers";
 import services from "./services";
@@ -12,6 +13,7 @@ export default {
   register,
   bootstrap,
   destroy,
+  config,
   routes,
   controllers,
   services,